requirements.txt
geoip2
pyramid
pyramid_jinja2
zope.component
zope.interface
gunicorn
